Move proc net tests into separate module
