+++ sip.cabal
@@ -41,6 +41,7 @@ library
 test-suite top-level-parser-tests
   type:               exitcode-stdio-1.0
   main-is:            Spec.hs
+  other-modules:      ProcNet
   hs-source-dirs:     test
   build-depends:
     base >= 4.7 && < 5
